# dicom-manager/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/job/info": {
            "get": {
                "description": "Список задач с пагинацией и фильтрами. Администратор может выбрать организации через organization_ids; без фильтра используется его организация. Остальные пользователи видят только собственные задачи. upload_source: manual — загрузка администратором через UI, clinic — загрузка специалистом через UI или поступление из Orthanc.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "job"
                ],
                "summary": "Получить информацию о задачах на обработку DICOM-файлов по их ID с пагинацией",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Offset для пагинации (по умолчанию 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit для пагинации (по умолчанию 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "collectionFormat": "csv",
                        "description": "ID организаций через запятую",
                        "name": "organization_ids",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "manual",
                                "clinic"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Источники загрузки DICOM через запятую",
                        "name": "upload_source",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о задачах успешно получена.",
                        "schema": {
                            "$ref": "#/definitions/get_paginated.GetInfoPaginatedOut"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/job/info/{job_id}": {
            "get": {
                "description": "Получить задачу по ID вместе с upload_source. При несовпадении organization_ids или upload_source возвращается 404.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "job"
                ],
                "summary": "Получить информацию о задаче на обработку DICOM-файла по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID задачи на обработку DICOM-файла",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "collectionFormat": "csv",
                        "description": "ID организаций через запятую",
                        "name": "organization_ids",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "manual",
                                "clinic"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Источники загрузки DICOM через запятую",
                        "name": "upload_source",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о задаче успешно получена.",
                        "schema": {
                            "$ref": "#/definitions/internal_api_job_get_by_id.GetByIDOut"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Задача не найдена.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/settings": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Параметры ротации организации (администратор)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.RotationSettings"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Сумма процентов не больше 100; центр × (1 + сумма / 100) не больше 8 мм. Готовые результаты не пересчитываются.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Сохранить параметры ротации для новых задач (администратор)",
                "parameters": [
                    {
                        "description": "Параметры ротации",
                        "name": "settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RotationSettings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.RotationSettings"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "get_paginated.GetInfoPaginatedOut": {
            "type": "object",
            "properties": {
                "jobs": {
                    "description": "Список информации о задачах на обработку DICOM-файлов.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.JobInfo"
                    }
                }
            }
        },
...
        "internal_api_job_get_by_id.GetByIDOut": {
            "type": "object",
            "properties": {
                "job": {
                    "description": "Содержит информацию о задаче на обработку DICOM-файлоа.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.JobInfo"
                        }
                    ]
                }
            }
        },
...
        "model.Decision": {
            "type": "string",
            "enum": [
                "approved",
                "rejected",
                "force_approved"
            ],
            "x-enum-comments": {
                "DecisionApprove": "Результат задачи принят специалистом.",
                "DecisionForceApproved": "Результат задачи принят специалистом, несмотря на рекомендацию системы.",
                "DecisionReject": "Результат задачи отклонен специалистом."
            },
            "x-enum-descriptions": [
                "Результат задачи принят специалистом.",
                "Результат задачи отклонен специалистом.",
                "Результат задачи принят специалистом, несмотря на рекомендацию системы."
            ],
            "x-enum-varnames": [
                "DecisionApprove",
                "DecisionReject",
                "DecisionForceApproved"
            ]
        },
...
        "model.JobInfo": {
            "type": "object",
            "properties": {
                "anatomical_region": {
                    "description": "Анатомическая область, к которой относится DICOM-файл.",
                    "type": "string"
                },
                "comment": {
                    "description": "Комментарий специалиста.",
                    "type": "string"
                },
                "confidence": {
                    "description": "Уровень уверенности в результатах обработки от 0 до 1.",
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "dicom_id": {
                    "description": "ID DICOM-файла.",
                    "type": "string"
                },
                "duration_ms": {
                    "description": "Время обработки задачи в миллисекундах.",
                    "type": "integer"
                },
                "error": {
                    "description": "Описание ошибки обработки.",
                    "type": "string"
                },
                "id": {
                    "description": "ID задачи на обработку.",
                    "type": "string"
                },
                "metadata": {
                    "description": "Дополнительные метаданные, связанные с задачей.",
                    "type": "object",
                    "additionalProperties": {}
                },
                "specialist_decision": {
                    "description": "Решение специалиста по результату задачи.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Decision"
                        }
                    ]
                },
                "specialist_id": {
                    "description": "ID специалиста, который принял решение.",
                    "type": "integer"
                },
                "status": {
                    "description": "Статус задачи на обработку.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.JobStatus"
                        }
                    ]
                },
                "updated_at": {
                    "type": "string"
                },
                "upload_source": {
                    "type": "string",
                    "enum": [
                        "unknown",
                        "manual",
                        "clinic"
                    ]
                },
                "violations": {
                    "description": "Список нарушений, обнаруженных в DICOM-файле.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "model.JobStatus": {
            "type": "string",
            "enum": [
                "pending",
                "processing",
                "completed",
                "failed"
            ],
            "x-enum-comments": {
                "JobStatusCompleted": "Задача успешно завершена.",
                "JobStatusFailed": "Задача завершилась с ошибкой.",
                "JobStatusPending": "Задача ожидает обработки.",
                "JobStatusProcessing": "Задача в процессе обработки."
            },
            "x-enum-descriptions": [
                "Задача ожидает обработки.",
                "Задача в процессе обработки.",
                "Задача успешно завершена.",
                "Задача завершилась с ошибкой."
            ],
            "x-enum-varnames": [
                "JobStatusPending",
                "JobStatusProcessing",
                "JobStatusCompleted",
                "JobStatusFailed"
            ]
        },
...
        "model.RotationSettings": {
            "type": "object",
            "properties": {
                "trochanter_center_mm": {
                    "type": "number",
                    "maximum": 8,
                    "minimum": 0.1,
                    "example": 2.7
                },
                "trochanter_tol_percent": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0,
                    "example": 63
                },
                "trochanter_yellow_percent": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0,
                    "example": 30
                }
            }
        },
...
    }
}